+++ app/src/main/java/com/sopt/now/compose/core_ui/component/textfield/TextFieldWithTitle.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose.component.textfield
+package com.sopt.now.compose.core_ui.component.textfield
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -13,7 +13,7 @@ import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
+import com.sopt.now.compose.core_ui.theme.NOWSOPTAndroidTheme
 
 @Composable
 fun TextFieldWithTitle(
